requirements.txt
connexion
fabric-comanage-api
flask-migrate
flask-sqlalchemy
psycopg2-binary
python-dateutil # alembic[tz]
python-dotenv
pytz
swagger-ui-bundle # connexion[swagger-ui]
pyjwt
#
# configparser
# cryptography
# uwsgi
